// whatsapp-clone-ui/src/app/services/core/models/chat-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Schema to hold the chat response
 */
export interface ChatResponse {

  /**
   * The chat profile image reference
   */
  chatImageReference?: string;

  /**
   * The bio of the chat
   */
  description?: string;

  /**
   * The chat Id
   */
  id?: string;

  /**
   * Flag to describe if the chat is group-chat
   */
  isGroupChat?: boolean;

  /**
   * Flag to determine if the another user in the single chat is online
   */
  isRecipientOnline?: boolean;

  /**
   * The last message sent in the chat
   */
  lastMessage?: string;

  /**
   * The last message time sent in the chat
   */
  lastMessageTime?: string;

  /**
   * The chat name
   */
  name?: string;

  /**
   * The list of other users ids in the chat
   */
  receiversId?: Array<string>;

  /**
   * The creator user id
   */
  senderId?: string;

  /**
   * The number of unread messages
   */
  unreadCount?: number;
}