// whatsapp-clone-ui/src/app/services/core/models/message-creation-resource.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Schema to hold the message creation data
 */
export interface MessageCreationResource {

  /**
   * The chat Id to add the message to it
   */
  chatId: string;

  /**
   * The message string content [Mandatory in case messageType=TEXT]
   */
  content?: string;

  /**
   * Flag to specify if the message is forwarded from another chat
   */
  isForwarded?: boolean;

  /**
   * List of media files
   */
  mediaFiles?: Array<Blob>;

  /**
   * The message content type
   */
  messageType: 'TEXT' | 'MEDIA' | 'AUDIO';

  /**
   * The message Id of the source replied message
   */
  repliedMessageId?: number;
}